add separate generators file (circ imps)
